JIT: more harmless restructuring
